requirements.txt
scikit-learn
numpy
matplotlib
Pillow